// src/lib/api/schema.ts
export interface components {
    schemas: {
// ...
        GeotrackResp: {
            /**
             * @description the longitude geo coordinate
             * @example 10.875482
             */
            longitude?: number;
            /**
             * @description the latitude geo coordinate
             * @example 49.514135
             */
            latitude?: number;
            /**
             * @description the accuracy of the position measurement in meter (radius)
             * @example 60
             */
            accuracy?: number;
            /**
             * @description the velocity in [m/s] between two points (this related to the last one in time)
             * @example 22.3
             */
            velocity?: number;
            /**
             * @description The time in [s] passed between two points (this related to the last one in time)
             * @example 11620
             */
            timediff?: number;
            /**
             * @description The distance in [m] passed between two points (this related to the last one in time)
             * @example 320
             */
            dist?: number;
            /**
             * @description identifier of the sender of this request
             * @example HASSIO
             */
            source?: string;
            /**
             * @description altitude of the position
             * @example 320
             */
            altitude?: number;
            /**
             * Format: date
             * @description date of this entry in ISO format
             * @example 2020-07-15T10:45:23.000Z
             */
            date?: string;
        };
// ...
    };
    responses: {
        /** @description Access token is missing or invalid; try to refresh your token */
        UnauthorizedError: {
            headers: {
                [name: string]: unknown;
            };
            content?: never;
        };
    };
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}